third_party/blink/web_tests/fast/xpath/4XPath/Core/test_literal_expr-expected.txt

PASS document.evaluate('""', document, null, XPathResult.STRING_TYPE, null).stringValue is ""
PASS document.evaluate('"Hi"', document, null, XPathResult.STRING_TYPE, null).stringValue is "Hi"
PASS document.evaluate('"NaN"', document, null, XPathResult.STRING_TYPE, null).stringValue is "NaN"
PASS document.evaluate('"• = middle dot"', document, null, XPathResult.STRING_TYPE, null).stringValue is "• = middle dot"
PASS document.evaluate('"0"', document, null, XPathResult.STRING_TYPE, null).stringValue is "0"
PASS document.evaluate('"1"', document, null, XPathResult.STRING_TYPE, null).stringValue is "1"
PASS document.evaluate('"2"', document, null, XPathResult.STRING_TYPE, null).stringValue is "2"
PASS document.evaluate('"3"', document, null, XPathResult.STRING_TYPE, null).stringValue is "3"
PASS document.evaluate('"4"', document, null, XPathResult.STRING_TYPE, null).stringValue is "4"
PASS document.evaluate('"5"', document, null, XPathResult.STRING_TYPE, null).stringValue is "5"
PASS document.evaluate('"31"', document, null, XPathResult.STRING_TYPE, null).stringValue is "31"
PASS document.evaluate('"-1"', document, null, XPathResult.STRING_TYPE, null).stringValue is "-1"
PASS document.evaluate('"-2"', document, null, XPathResult.STRING_TYPE, null).stringValue is "-2"
PASS document.evaluate('"-3"', document, null, XPathResult.STRING_TYPE, null).stringValue is "-3"
PASS document.evaluate('"-4"', document, null, XPathResult.STRING_TYPE, null).stringValue is "-4"
PASS document.evaluate('"-5"', document, null, XPathResult.STRING_TYPE, null).stringValue is "-5"
PASS document.evaluate('"3.1415926535"', document, null, XPathResult.STRING_TYPE, null).stringValue is "3.1415926535"
PASS document.evaluate('0', document, null, XPathResult.NUMBER_TYPE, null).numberValue is 0.0
PASS document.evaluate('0.5', document, null, XPathResult.NUMBER_TYPE, null).numberValue is 0.5
PASS document.evaluate('-0.5', document, null, XPathResult.NUMBER_TYPE, null).numberValue is -0.5
PASS document.evaluate('1', document, null, XPathResult.NUMBER_TYPE, null).numberValue is 1.0
PASS document.evaluate('-1', document, null, XPathResult.NUMBER_TYPE, null).numberValue is -1.0
PASS document.evaluate('1.5', document, null, XPathResult.NUMBER_TYPE, null).numberValue is 1.5
PASS document.evaluate('-1.5', document, null, XPathResult.NUMBER_TYPE, null).numberValue is -1.5
PASS document.evaluate('2', document, null, XPathResult.NUMBER_TYPE, null).numberValue is 2.0
PASS document.evaluate('-2', document, null, XPathResult.NUMBER_TYPE, null).numberValue is -2.0
PASS document.evaluate('2.6', document, null, XPathResult.NUMBER_TYPE, null).numberValue is 2.6
PASS document.evaluate('-2.6', document, null, XPathResult.NUMBER_TYPE, null).numberValue is -2.6
PASS document.evaluate('3', document, null, XPathResult.NUMBER_TYPE, null).numberValue is 3.0
PASS document.evaluate('-3', document, null, XPathResult.NUMBER_TYPE, null).numberValue is -3.0
PASS document.evaluate('31', document, null, XPathResult.NUMBER_TYPE, null).numberValue is 31.0
PASS document.evaluate('4', document, null, XPathResult.NUMBER_TYPE, null).numberValue is 4.0
PASS document.evaluate('-4', document, null, XPathResult.NUMBER_TYPE, null).numberValue is -4.0
PASS document.evaluate('4.5', document, null, XPathResult.NUMBER_TYPE, null).numberValue is 4.5
PASS document.evaluate('-4.5', document, null, XPathResult.NUMBER_TYPE, null).numberValue is -4.5
PASS document.evaluate('5', document, null, XPathResult.NUMBER_TYPE, null).numberValue is 5.0
PASS document.evaluate('-5', document, null, XPathResult.NUMBER_TYPE, null).numberValue is -5.0
PASS document.evaluate('-42', document, null, XPathResult.NUMBER_TYPE, null).numberValue is -42.0
PASS successfullyParsed is true

TEST COMPLETE
